Applied Andrea Arcangeli <andrea@suse.de> patch. Note below:
I was used to get reports of nanosleep02 failing on -aa, so I had a look
and as expected they were false positives. So I fixed the problems in
the testcase, and now I verify also for a good precision of the order of
the usec, to ensure that no userspace live lock can happen under a flood
of signals (i.e. with glibc). The live lock can happen with a low
precision nanosleep rem field, because to be correct it has to
approximate the remaining time for excess, so with mainline it would
keep returning that 10msec are missing (or 1msec with 2.4) and one can
get 1 signal every msec generating the userspace live lock in nanosleep
restart.

So I'm afraid the new testcase will work successfully only on -aa based
kernels. To allow mainline to pass the test too, increasing
USEC_PRECISION to 11000 (11msec) should be enough I guess, I didn't
tested it. but I wouldn't recommend to increase the allowed usec
precision over 100usec to catch implementations that can introduce
userspace live locks under heavy signal handling (again, of the order of
one signal every 10msec, or 1msec with 2.5).