Added calls to tst_tmpdir & tst_rmdir. Alos applied patch from Andi Kleen:
clone expects the new value of the stack register as child_stack
argument. On an architecture with up-to-down stack this
has to be the end of the stack area.

Previously it was passing the beginning.

At least when you're not on PA-RISC this patch is correct.
On PA-RISC the stack grows up so the old code was correct.
For everybody else this change is needed. PA-RISC will need to be
separately addressed.

On x86-64 it would corrupt enough memory to cause a segfault
in free. It is a mistery why it worked on i386 or wherever
it was tested.

-Andi