Applied patch from Jacky Malcles: -------------------- according to the testcase code and to the documentation,
 Syscall specifications of Linux of semop:
...
cmd is lower than zero: if the value of the semaphore is greater than
the
absolute value of cmd, the value of the semaphore is
decremented by the value of cmd and the call returns. Otherwise, the
call
block until: the value of the semaphore becomes greater than
the absolute value of cmd at which time the value of the semaphore is
decremented by the value of cmd and then the call returns, the
semaphore set is removed, or a signal occurs. The calling task must have
alter privileges on the semaphore set.
...etc...

so,
I simply added an initialization to the semaphore.