warnings cleanup AND Dan Kegel's patch: I suspect the standard allows functions like read() to return 0 if no data was transferred even if the buffer pointer was invalid.  (Seems logical, because it's only when preparing to transfer data that the kernel might notice the bad address.) See http://www.geocrawler.com/archives/3/35/1996/10/0/222307/ for language backing this up.
recv01, recvfrom01, and recvmsg01 all fail most of the
time on some platforms because a race condition
in the test causes no data to be ready at the time
of the recv(), causing the expected EFAULT to not appear.

The attached patch adds a select() call to recv01.c to ensure
that data is ready at the time of the recv().
It also fixes a number of compiler warnings.
If y'all like the patch, I can do a similar one for
recvfrom01 and recvmsg01.
- Dan