Applied Dan Kegel's patch: As explained in a message earlier today, nanosleep01 is racy; it will fail if the time rolls over to a new second during the time between the three crucial syscalls. This happens much more often than you'd suspect on some platforms.
Attached is a patch that replaces calls to time() with
calls to gettimeofday().  The tolerance to which the
nanosleep duration is measured is given by the constant
SLOP_MS.  On a modern PC, the slop is on the order of 17ms.
I have set SLOP_MS to 250ms to be safe on slower platforms.
- Dan