Applied patch from Susanne Wintenberger: <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<< the source code  issues a  'cat /proc/swaps | wc -l' to get all swapfiles (to get all lines of /proc/swaps), if there are any.  Afterwards the number of swapfiles to be added is calculated in the following way:
swapfile = MAX_SWAPFILES - swapfile + 1;
number of swapfiles to be added = 32 - current swapfiles + 1;

The reason, why the calculation adds 1 at the end is, that the /proc/swaps file includes a header line, which is 'included in swapfiles'. This works fine for 2.4 and as soon as there is always a header line, even no swapfiles exist. But the display of /proc/swaps/ was completely changed to 2.5. Now if there is no swapfile, no header line is displayed. This leads to a 'swapfile' calculation error, since now the program tries to create 33 swapfiles, which is not possible, and which results in system error messages.
as a fix I added the following lines to the code
        if (swapfile > MAX_SWAPFILES) {
             swapfile = MAX_SWAPFILES;
        }
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>