Applied Matthew Wilcox flock03 patch: <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<< The reason flock03 currently fails on both Linux 2.4 and 2.5 is because it's wrong ;-)  I suspect the test author didn't understand flock locks -- they're very different from POSIX locks.  flocks are associated with a file descriptor, but the child process was opening a new file descriptor and expecting to be able to unlock it.  That won't work; you have to pass the file descriptor to the child from the parent.
This patch does just that.  I'm not sure it's exactly what the author
wanted to achieve.  I've added the flock06 test too which tests that
flocks on two fds on the same file do conflict with each other.

flock03 still fails.  This is due to a real bug in both Linux 2.4 and 2.5.
Andrea's kernel picked up the fix, but Marcelo has declined it.  I hope
to fix it properly in 2.5 in the next couple of weeks.
>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>