Finally got the correct resolution on the POSIX issue from Mark Brown: <<<<<<<<<<< Hello-
The current POSIX standard (IEEE POSIX 1003.1-2001) (which is also the Open Group
Single UNIX Specification v3, quoted below) currently states the following. This is
the text as it appears after taking into account Technical Corrigendum 1:

The pthread_detach( ) function shall fail if:
[EINVAL] The implementation has detected that the value specified by thread does not
refer to a joinable thread.
[ESRCH] No thread could be found corresponding to that specified by the given thread
ID.
The pthread_detach( ) function shall not return an error code of [EINTR].

"shall fail" means "must detect the error condition in question and return the error result".

This is going to be changed in the not-yet-released Technical Corrigendum 2 to read
"may fail".

"may fail" means that an application should not depend on the error being detected
or returned as an error.

Some implementors already know about the TC2, and may have taken action on it.

-------------------
Mark S. Brown
STSM, UNIX/Linux OS Standards
IBM Systems Group
bmark@us.ibm.com
>>>>>>>>>>>>