These tests were receiving strange SIGSEGV calls on ppc64 when compiled into 64bit.  Further investigation showed that these tests were using the incorrect system call to terminate the children, return().  As documented in the man page, the clone() system call expects its children to use exit().  The ppc64 64-bit kernel was simply catching this, while other kernels, i.e. i386, were handling it.  I corrected this in all the tests.
clone02 - This test was having trouble cleaning up on ppc64 64-bit.  I'm
not sure if it's a problem with our tst_sig API or the Red Hat 2.4 kernel
I was testing with.  Anyhow, it was easier to remove the tst_rmdir() and
replace with simple remove() in this case.

clone04 - This test was also bad b/c it was testing for EINVAL in 2 ways,
when there is only one supported/documented way to get EINVAL.  I removed
the other case.