Applied patch from Erik Andersen: ================================ The string tests check strncmp(), but they expect certain implementation specific results, which are not correct.
SuSv3 states:

		 "Upon successful completion, strncmp() shall return an
		 integer greater than, equal to, or less than 0, if the
		 possibly null-terminated array pointed to by s1 is
		 greater than, equal to, or less than the possibly
		 null-terminated array pointed to by s2 respectively."

yet the current strncmp() test checks for the result for equality
with things such as ('y'-'a'), which assumes a particular
implementation that is much more specific than is authorized by
SuSv3.  I think the test should merely check for >, ==, or <
zero, and nothing else.

I propose this patch be applied to remedy the strncmp test, and
bring it into exact compliance with SuSv3....
================================