Applied patch from Ming Gao: ============================ When we run LTP-20031204 in RHEL 3 Update 1, tcpdump01 and  mc_cmds also failed. We think they have the similar cause with the test case. This problem is about multiple network interfaces.
The following revisions added support for multiple interfaces:

http://cvs.sourceforge.net/viewcvs.py/ltp/ltp/testcases/network/tcp_cmds/tcpdump/tcpdump01?r1=1.6&r2=1.7
http://cvs.sourceforge.net/viewcvs.py/ltp/ltp/testcases/network/multicast/mc_cmds/mc_cmds?r1=1.4&r2=1.5

However, I thought these revisions might not be the best solutions.

My recommendation is: We can export the value of IFNAME and INTERFACE when we setup the test environment.
For example:

export INTERFACE=172.30.1.70
export IFNAME=eth1

At the same time, use the following patch to the latest version of tcpdump01 and mc_cmds.
=============================