Applied patch from Erik Andersen: ================================ I just got around to running the latest ltp with uClibc and noticed a new FAIL regression in pipe07.  After hunting through uClibc's stdio code I was puzzled to find we were internally trying to fopen a filename with pipes and redirects in it, rather than the filename that had been explicitly passed into fopen. The filename matched up with the format string from pipe07.c line 87.  The made me strongly suspect a buffer overflow.  Running 'valgrind' on pipe07 confirmed that uClibc's _stdio_write was doing some scribbling where it shouldn't be.
Visual inspection by myself and my colleague Manuel Novoa quickly
led us to notice the actual buffer overflow, which occurs on line
87-88 of pipe07.c when printf writes into a malloc'd buffer the
size of a char pointer (i.e. 4 bytes on x86).  The rest of the
string was then scribbled all over.  This bug is a new regression
introduced within the last month, and certainly explains the "For
some reason, a SIGSEGV was generated" comment in rev 1.4.

Please apply,

 -Erik
================================