Applied patch from Andi Kleen: ============================== While running LTP on x86-64 I often get hangs in the EFAULT test in syslog12.
The reason is that it tries to test for the EFAULT handling by passing an
illegal address to syslog(2, ...). The kernel then first waits for an
event in the log buffer before copying out the data and possibly noticing
the EFAULT condition. But when the kernel log buffer is empty and nothing
logs a new event the test hangs forever.

This patch just adds a 1s timeout to at least prevent the hang.

Better would be probably to just remove this one because I cannot see how
it should be made to work reliably.

-Andi
==============================