Applied patch from Gary Williams: ================================ Alright, a couple of the changes in here are simply because I like a test to say something other than "done".
But, one change fixes how the return value from pthread_create is used.
Pthread_create does NOT set errno.  If there is an error, it returns the
error value.  This is documented in the man page.  Pthread_create not
touching errno can be seen by looking at the source code (or by talking
to your local BSP developer, if you'd rather).

Anyway the file in question is:
    testcases/kernel/sched/pthreads/pth_str02.c


Gary
================================