Applied patch from Andrea Arcangeli: ========================== I've got a bugreport showing mmap failures during mtest05 mmstress.
It's clearly not a kernel bug, and to me it looks like it wasn't
intentional to leak all the mmap mappings since you're deleting the
files after the test. the -t 1 parameter will always result in memory
allocation failures in normal hardware, so I wrote a patch to fix the
memory leak.
==========================