Applied patch from Gao, Ming: ==================================================== We found some problems about network testcases in ltp.
These problems happen when we have a machine whose hostname is a sub-string of other machine's hostname. For example, we have a machine named "plinuxt9" and another machine is "plinuxt9b". In case rwho01, when case goes to:

      rwho -a > $OUTFILE
      HOST=`awk '{print $2}' $OUTFILE|grep $LHOST|cut -f1 -d:|sort -u`
      [ "$HOST" = "$LHOST" ] || end_testcase "$LHOST is not in rwho outfile"

If we run this on plinuxt9(that means $LHOST = plinuxt9), HOST will gets a value:
plinuxt9
plinuxt9b

not a expected value:
plinuxt9

This will make the case fail.

The solution is to change the way of grep. We can change the second line of above code to:

      HOST=`awk '{print $2}' $OUTFILE|grep "$LHOST\>"|cut -f1 -d:|sort -u`

Then we will get what we expected.

I found some other cases have similar problem and create a patch for each of them for ltp-full-20040506.
====================================================