Applied patch from Brian Reitz: ===============================
I'm attaching a simple patch that fixes this issue.  My apologies if
this is the wrong list to post these sort of issues to.  If it is easy
to make these changes (create a bug, submit a fix) myself I'm willing to
do it if somebody can help me understand the steps I need to take.

I don't subscribe to the ltp-list so anything you post to that list that
you want me to reply to will require a copy sent to me directly (sorry).

Thanks,
Brian Reitz