Applied patch from David Marlin: I am working with the 20040804 release of LTP.  I noticed that the tcpdump01 test was failing for me because IFNAME was undefined.  I could not find where it was supposed to be set in tcpdump01, so I added a line to set IFNAME if not already defined, modeled after code in 'mc_cmds'.
Attached is a patch showing the code I added.  Note: there are two NICs
in the system I am testing.  I assume that the first (eth0) should be
used as the interface for this test, unless otherwise specified.

Please let me know if IFNAME should be set using some other method, or
if this is an acceptable approach.


Thank you,

d.marlin