Applied patch from David Miller: ====================================================================== fcntl16.c (and probably a bunch of other places in the ltp testsuite) use the following scheme:
static int alarm_flag;

static void catch_alarm(int signo)
{
		 alarm_flag = 1;
}

		 signal(SIGALRM, catch_alarm);

		 ...

		 alarm(TIME_OUT);
		 wait(&status);

Just a quick scan shows that, among the fcntl tests alone,
this technique is used in fcntl14.c, fcntl15.c, fcntl16.c,
and fcntl17.c

The expectation is that if no wait() events occur, the SIGALRM
will cause the wait() to return with -EINTR after the signal
handler catch_alarm() runs.

This never happens, because signal() gives BSD signal semantics
and part of that is that the SA_RESTART flag is passed into the
actual system call (usually rt_sigaction) used to register the
signal handler.

SA_RESTART means that the system call is restarted, not returned
from, when the signal handler returns.
In order to fix this, test cases using this technique in LTP
will need to explicitly call sigaction() with the appropriate
flags.  In particular, it will need to have the SA_RESTART
flag clear in such calls.

What's really fascinating to me is that this test is passing
for somebody out there, I can't believe I'm the only person
hitting this :-)

I ran my little test program above both on x86 and sparc64, just
in case it was some difference in glibc's signal() implementation,
but both cases pass the SA_RESTART flag in and both cases restart
the wait4() call after the signal handler runs and the program thus
hangs forever afterwards.

I note that recently there were some "portability" changes for AIX
and Solaris that tried to start using sigaction().  These changes
are buggy because although they clear sa_flags the rest of the
structure besides sa_flags and sa_handler is not initialized.  A
memset() is needed on the whole sigaction structure, and also a
sigaddset(&act.sa_mask, X) is necessary (where X is the signal
being registered) before initializing the other elements.

But once again, I have no idea how these ltp tests work for other people
and this bug has been present in ltp for a very long time.

Any insight into that matter would be very appreciated. :-)