Applied patch from Mitsuru Chinen: Hi all,
rlogin01 and telnet01 tests check the connectivity of each utility
based on the output `ls -l /etc/hosts | wc -w'.
They assumes the output is `9'.

However, the output is changed when locale was changed.

  $ LC_ALL=C ls -l /etc/hosts | wc -w
  9

  $ LC_ALL=en_US.UTF-8 ls -l /etc/hosts | wc -w
  8

So, I proposed to modify the test script to run the command line
always on C locale. The following is my proposed patch.