Applied patch from Johnny Klonaris: =================================== In our testing we discovered that if the output of runltp is piped into another task the parent shell would hang after the tests completed.
I tracked this to a stray copy of the kill02 test being left behind in
a separate process group, attached to init. The parent copy of kill02
was quitting and causing a SIGTERM to be sent to the "1" copy before it
could terminate the "B" copy (the one left behind).  The attached patch
adds a wait(2) to the par_kill function causing it to wait until signalled
tasks complete.

-jk