Corrected a logical typo in the test found by John Clemens: While tracking down random SEGV's in the mmapstress test suite, I think there's an typo in mmapstress07.c:
lines 199-200:
         if ((mapaddr = mmap((caddr_t)0, pagesize*2 + holesize, PROT_READ,
                 MAP_SHARED|MAP_FILE, rofd, sparseoff)) == (caddr_t)-1) {

we mmap holesize plus 2 pages (in our test case, 4096, or one page.. for a
total of 3 pages).  we then do some stuff, none of which involves mmap, or
mremap, and then we unmap the above space thus:

line 244:
         if (munmap(mapaddr, holesize + 3*pagesize) == -1) {

Errr... 3*pagesize? that means we'd unmap 4 pages, not 3 like we mmaped in
the above example..  typo? bug? or am I missing a nuance?  Seems to me
this should be 2*pagesize.

This was causing segfaults for us because when compiled dynamicaly, it
just so happens that the application itself (or one of the shared libs)
was mapped directly after this mmap'd area, so we blew away its first
mapped page.