Marc Tardif writes: When getting the ltp to run on sparc, I came across the following error:
  shmt09      1  PASS  :  sbrk, sbrk, shmget, shmat
  shmt09      2  PASS  :  sbrk, shmat
  shmat1: Invalid argument
  shmt09      3  FAIL  :  Error: shmat Failed, shmid = 294916, errno = 29

The reason for this failure is because the shmat system call is being
passed an address which cannot be rounded down to the nearest multiple
of SHMLBA. To demonstrate, on sparc, the following address is being
calculated:

  vp = (void *) ((char *)sbrk(0) + 4 * K_1);

Considering K_1 is a macro defined as 1024, the pointer passed to shmat is
only 4096 bytes above the current heap address whereas the SHMLBA macro
would require this address to be at least 8192 bytes above. I say at least
because proper alignment must also be taken into account. So, there are
really two problems here:

First, the value of SHMLBA is being assumed in the source code and only
exceptionally set for __mips__ and __ia64__. Instead, I would suggest using
the actual value of SHMLBA to compute a valid pointer address.

Second, the value of the pointer address passed to shmat should be set in
such a way that rounding down to the nearest multiple of SHMLBA results
above the current heap.

To solve both issues, I have attached a patch which I have successfully
tested on sparc and i386 architectures.