Ssem_post/5.-1.c According to its intention, #3, call sleep(1) and then alarm(1).  Moving sleep before alarm causes the test to pass
sem_post/8-1.c
Although step 3 intended,  The children lock the semaphore. * Make sure the two children are waiting."  Without that caeratn piece of cde which explicity make children wat.  the test will PASS and the children are waiting.

Sem_unlink_1_1 and sem_unline_2_1
"Sem_unlink" is too long for the name of a semaphore for certain architectures