I spent some time looking at the sched_latency test trying to see why it took so long to start the thread.  The java version of the test takes a comparable amount of time to start the thread as it does for each period. The rt-test version was taking an order of magnitude longer.  After digging through the java documentation, I found that the RealtimeThread does not actually start executing until the RealtimeThread.waitForNextPeriod() fulfills the time specified in the PeriodicParameters parameter included in its constructor.  So all the tests are scheduled to run about 1 second after the threads are created.  This gives the system plenty of time to create the thread (a non-realtime code path), set the scheduling parameters and then sleep until the specified run time.  The start latency as measured by the test goes something like this:
start_time = nanoTime() + 1000000000;
rt = new realtimeThread(sched_param, new \
        PeriodicParameters(start_time, ...), ..., run);

/* by the rtsj specification, this method will not run until start_time */
run() {
        now = nanoTime();
        start_latency = now - start_time;
        ...
}

The rt-tests version went something like this:

start_time = rt_gettime();
create_fifo_thread(periodic_thread, (void*)0, PRIO);

periodic_thread() {
        now = rt_gettime();
        start_latency = now - start;
        ...
}

As you can see, the start_latency is a very different measurement in these two cases.  In the first, it is basically the initial starting latency as measured by the remainder of the loops.  In the second, the execution path of creating the pthread is also counted in the latency.  While this might be an important number for us to keep low in our testing, this is not what the java test is measuring. This patch puts the rt-test version in line with the java version.  I also think this is reasonable since the rt-wiki says that we should _NEVER_ create threads on the real-time stage. Signed-off-by: Vernon Mauery <vernux@us.ibm.com>.