This patchset creates the basic infrastructure for running the existing network testcases under containers automatically. The patchset creates a command file named nw_under_ns for the purpose. Please review them and provide all your valuable feedback. I have addressed all the comments from lxc-dev list except one. The changelog looks like: 1. Added verbose comments and patch description. 2. Restructured the licensing text. 3. Merged two patches into one touching runallnetworktest_child.sh file.
TODO:
1. Changing the kernel version to proper one
2. unshare not needed as clone only is suffucient. I will do it in a separate patch as need to do extensive testing to ensure other testcases do not break.

[patch 03/12] CONTAINERS change the way child pid is received:
This patch changes the way information is exchanged between parent and child. This is required because of the following reason: Earlier the child was in the same pid namespace as parent and this info was passed from the child to the parent through a fifo. But now the child is in a separate pid namespace and hence this pid will not make any sence to the parent.
Signed-off-by: Sudhir Kumar <skumar@linux.vnet.ibm.com>, Acked-By: "Serge E. Hallyn" <serue@us.ibm.com>.