1. I screwed up and broke install by accident when I was monkeying around with env_post.mk. 2. Make in-build-tree testing a little more conducive as per Mike F's convincing argument:
requiring install-related variables to be sane in order to do `make install`
is fine (i.e. prefix/etc...).  but that is the only time install related
variables should be needed/consulted.

when you compile a package (i.e. run `make`) regardless of this being done at
the toplevel or a subdir only, the build process should never need anything
ltp related outside of the source tree (so i'm not talking about libcap or
whatever, just libltp).  this is the expected behavior for all packages out
there, not just ltp.

one reason for why this is a bad idea is that it is then impossible to
compile/test a new version of a package locally before installing it into the
live filesystem.  you would have to first uninstall the old version (rm -rf
/opt/ltp), then compile/install the new version, and then hope that (1) the
build is successful as well as (2) the package working as expected (bug free).
-mike

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>