realtime: Remove printing of cpu affinity masks in pi-tests: Some tests under realtime/func/pi-tests in LTP display junk values for cpu affinity masks: Start ./testpi-2 protocol in mutexattr is 1 Thread 14574 started running with priority 10 on CPU 1110536304 Thread 14574 at start pthread pol 2 pri 10 - Got global lock Thread 14575 started running with prio 20 on CPU 1084272752 Thread 14576 started running with prio 30 on CPU 1118929008 Thread 14577 started running with prio 40 on CPU 1127321712 Noise Thread 14578 started running with prio 40 on CPU 1099915376
testpi-1 and testpi-4 also report similar wrong values.

The cause of this problem is the incorrect way in which sched_getaffinity output is used in the tests. I think there is no need to call getaffinity() to know the cpu mask, because we test the return value of setaffinity() before that. The following patch removes all calls to sched_affinity() and removes printing cpu affinity masks.
Signed-off-by: Sripathi Kodi <sripathik@in.ibm.com>,
Acked-by: Gowrishankar <gowrishankar.m@in.ibm.com>,