drop adding of -O/-W type flags that are already handled in common .mk files
for the few makefiles that arent yet converted to the .mk infrastructure:
Lately gcc developers introduced -Wextra flag that does the same as -W but is
more descriptive. According to this using -W flag should be safe (as gcc is
backward compatlible) but using -Wextra is limited to newer gcc releases.

Attached patch replaces all -Wextra occurences with -W and thus fixes
compliation failures with older gcc (mine was gcc-3.3.3 on sles).

patch by Cyril Hrubis