Hello guys,
I have run into some problems with the network stress tests. I submitted
a bug through the sourceforge bug tracker, but after subscribing to the
mailing list I can see that it's here where all the action is. If this
is not the right place, then please tell me what is the preferred way to
submit bug reports and patches.

I am using ltp-full-20091031.

A lot of the TCP tests involving packet loss fail with this error:

./tcp4-uni-basic01: line 394: netem_param: unbound variable

The reason is that strict checking for undefined variables is enabled
with "set -u", but the variable netem_param is not always defined before
testing its contents.

The fix I have applied is to initialize the netem_param variable with an
empty string.

Also if4-addr-change always fails with this error:

TBROK  :  Failed to set an IPv4 address at the remote host

I have found the reason to be a missing semicolon.

Best regards,
Radoslav Kolev

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>