Hello,
I tried to make ltp with the -j switch and found the following problems:

 - the autotools target fail because of missing dependency on aclocal.m4 (almost all autotools targets depend on it)

 - the shell loop in generic_trunk_target makes the "all" rules of the subdirectories run strictly sequentially

This patch fixes these issues by adding dependencies to the autotools targets (it also deletes some of the duplicated targets from the realtime Makefile), and by making one target for each subdirectory in generic_trunk_target.

Note that I really don't know if this works on an arbitrary make version.

Regards
   Jiri Palcek

Signed-off-by: Jiri Palecek <jpalecek@web.de>