Hello,
I've found some problems with the sendfile01 testcase (the one under network/tcp_cmds). This patch fixes them:

 Pass the -6 parameter to gethost if we are to use ipv6 and run the
 ipv6 server in that case. Also, fix the awk code for getting the PID:
 The PID is the second column, which is $2, not $1.

Also:

 - initialize addrlen for accept syscall in the server (=> avoid EINVAL)

 - use tst_resm in the client instead of tst_res, to avoid an error message interpreted as a filename

Regards
   Jiri Palecek

Signed-off-by: Jiri Palecek <jpalecek@web.de>

The patch originally sent via email was modified so that the POSIX form of ps was called (with the -), and the PID was in the first column.
Furthermore, this patch was modified to ensure that IPADDR was in fact being set to a valid value to avoid hangs as gethost could be missing (as well as netlib.sh), which symbolizes an install failure, which would ultimately result in a test hang..