Running the ltp suite on ChromeOS (x86 with a 2.6.32 kernel) fails linkat01 testcase 21:
linkat01   21  TFAIL  :  linkat() failed: TEST_ERRNO=EXDEV(18): Invalid
cross-device link

This is probably because my /tmp is mounted from someplace other than the
filesystem that contains newdirfd.
The goal of testcase 21 is to demonstrate that linking a directory will fail
EPERM.  The fix is to use olddir/. instead of /tmp.

Tested-by: Henry Yei <hyei@mvista.com>
Signed-off-by: Bryan Freed <bfreed@google.com>
Signed-off-by: Rishikesh K Rajak <risrajak@linux.vnet.ibm.com>