The test_fs_bind testscript hangs indefinitely when run on a machine where busybox applets are used.
The test hangs while trying to restore the mounts after a test run.

This happens because the "xargs" applet from busybox doesn't understand
the --max-args=1 option. Thus, the mounts from the test are not removed
and this goes on in a loop, indefinitely.

This could be avoided by using "-n" option which is supported by both
the implementations of xargs.

Signed-off-by: Suzuki K P <suzuki@in.ibm.com>