Divide and conquer Makefiles to for gmake/pmake.
The problem with the previous Makefiles is that it worked on pmake, but
not gmake, as the previous Makefiles made use of a non-POSIX secondary
expansion extension in pmake (gmake uses the .SECONDEXPANSION keyword to
achieve the same thing). So, make this completely POSIX.

Still doesn't fix the phantom targets issue. I need to start digging at
the locate-tests level.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>