[PATCH v4] Remove files which does not have a dmesg entry
The recently introduced option:

-K DMESG_LOG_DIR
        Log Kernel messages generated for each test cases inside this directory,

when used will generate huge no. of files inside the DMESG_DIR. Many of them will be ZERO size as we do not expect KERNEL MESSAGE to be generated against all the LTP TESTS that we will run. It is better to prune them, and let only those files exist which will have real entries. The following patch fixes the same.
Signed-off-by: Subrata Modak <subrata@linux.vnet.ibm.com>,