Improve output and notify of test naming issues.
The output now better matches the previous code (but not exactly because
I've simplified the problem into compile, instead of cc/ld failures for
simplicity in the make rules).

Test naming is wrong for a boatload of tests (some tests are buildonly,
but it turns out they have a main so they can be run (and most of the
case) the other way around.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>