Improve logic breakdown and flow. Fix bugs.
The overall description for generate_makefile didn't match the behavior;
isolate the Makefile producing portion, so the code could be better
tested by itself, and the flow becomes more intuitive.

Oh yeah, and fix $AUTHORDATE to print out less crap, and don't print out
the Makefile if Makefile.2 (the targets section) does not exist.

Also, punt out the last cached data in generate_makefiles, as we were
missing the last Makefile (and thus... it was regenerating the Makefiles
multiple times unnecessarily at the top-level Makefile, with GNU make in
particular, compared to BSD make -- which was bizarre).

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>