Fix oversight with `test tools'.
All of the *core.c files are used as test tools for the *.sh scripts
when running tests. Previously, they were being leaked into the test
execution, which was wrong, so I had to come up with an enhanced scheme
for detecting what needs to be compiled as well as what needs to be
tested.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>