Be pedantic about the ret. code from timer_create.
Opengroup says:

If the call succeeds, timer_create() returns zero and updates the
location referenced by timerid to a timer_t, which can be passed to the
per-process timer calls. If an error occurs, the function returns a
value of -1 and sets errno to indicate the error. The value of timerid
is undefined if an error occurs.

So, technically non-zero != -1 is undefined behavior.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>