Use better precision accounting for pselect01.
In my recent tests, not all desired time is equal to exact execution time,
sometimes it is +/-1s, and in a few times it becomes +/-2s. In most cases,
+/-2s is still a valid value between exact execute time and desired time.

(the original patch proposed using time(3), but due to loss in precision
as time(3) specifies in quantum of seconds, clock_gettime(3) was recommended
and subsequently implemented).

Signed-off-by: Caspar Zhang <czhang@redhat.com>
Acked-by: Garrett Cooper <yanegomi@gmail.com>