Don't assume that SHELL has a sane value.
SHELL is set to the login shell for the user (should have verified that
before), and on FreeBSD the default login shell is tcsh (bleh). So take
tcsh out of the equation by hardcoding sh -- which is ash on Debian variants
and bash on almost all other Linuxes -- to increase determinism.

Signed-off-by: Garrett Cooper <yanegomi@gmail.com>