drop test pidns03 ? I am not sure about the original intent of pidns03 test case but I think the test is expected to fail :-( The header for this test says:         * Verify that:         * 1. When parent, clone a process with flag CLONE_NEWPID, see the         * process id of the parent is existing after mounting /proc
The test passes in the parent-pid from init-pid-ns and looks for that pid in the child's mount of /proc. But for the CLONE_NEWPID child, parent process should appear to have a pid == 0. The pid of the parent in the initial pid namespace must NOT be visible to the child - in fact a different process can have that pid.
I am not surprised the test fails - I am curious to know the version of the kernel when this test reliably passed - if at all. I tried this on 2.6.29 and 2.6.34-rc5 and it fails reliably.
I dug through my history and see that this test was part of a 6-patch set, We did have some comments on the coding style on the set, but we seem to have missed the intent of this test :-(.
Do we have any other history of this test ? If not, I think we should either drop the test or modify the test to verify the opposite (i.e.  verify that the opendir() in the test fails with ENOENT). Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>, Signed-off-by: Subrata Modak <subrata@linux.vnet.ibm.com>,